Add Capability column for new Marketplace filtering
Per the conversation of the TG-Content group from 30JLY2020,
this PR adds the Capability column for the Cores and SoC tables
to allow selection by RV32, RV64, or RV128 (no entries match
RV128). Some of the entires support both RV32 and RV64, which I
added by comma separating the two values in the column. I do not
know if the search plugin can handle that, if it cannot please
let me know and I will duplicate the rows for both capabilities.